\name{DSpat.lines}
\alias{DSpat.lines}
\docType{data}
\title{Example DSpat lines dataframe}
\description{ An example dataframe of 10 transect centerlines}
\usage{data(DSpat.lines)}
\format{
  A data frame with 10 observations on the following 5 variables.
  \describe{
    \item{\code{label}}{unique line label that links points to lines}
    \item{\code{x0}}{x-coordinate for beginning of line}
    \item{\code{x1}}{x-coordinate for end of line}
    \item{\code{y0}}{y-coordinate for beginning of line}
    \item{\code{y1}}{y-coordinate for end of line}
  }
}
\details{
An example set of vertical lines evenly spaced across a window of 100x100.
}
\keyword{datasets}
